[#45]refactor: revert
